# docs/swagger.yaml
basePath: /
definitions:
  api.CalculationResponse:
    properties:
      calculation:
        type: string
      result: {}
      session_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
      warnings:
        items:
          type: string
        type: array
    type: object
  api.CalculatorInfoResponse:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      success:
        type: boolean
    type: object
  api.FieldDiff:
    properties:
      description:
        type: string
      difference:
        type: number
      field:
        type: string
      new_value: {}
      old_value: {}
    type: object
  api.HealthResponse:
    properties:
      build_time:
        example: "2024-01-01T12:00:00Z"
        type: string
      status:
        example: healthy
        type: string
      time:
        example: "2024-01-01T12:00:00Z"
        type: string
      version:
        example: 1.0.0
        type: string
    type: object
  api.SolverComparisonReport:
    properties:
      converged_new:
        type: boolean
      converged_old:
        type: boolean
      error_new_vs_old:
        type: number
      error_new_vs_theory:
        type: number
      error_old_vs_theory:
        type: number
      field_differences:
        items:
          $ref: '#/definitions/api.FieldDiff'
        type: array
      final_value_new:
        type: number
      final_value_old:
        type: number
      final_value_theory:
        type: number
      improvement_ratio:
        type: number
      initial_value:
        type: number
      issue_summary:
        items:
          type: string
        type: array
      iterations_new:
        type: integer
      iterations_old:
        type: integer
      max_error_new:
        type: number
      max_error_old:
        type: number
      mean_error_new:
        type: number
      mean_error_old:
        type: number
      recommendations:
        items:
          type: string
        type: array
      test_equation:
        type: string
      time_points_match:
        type: boolean
      time_range:
        type: number
      time_step:
        type: number
    type: object
  api.SolverComparisonResponse:
    properties:
      calculation:
        type: string
      comparison_report:
        $ref: '#/definitions/api.SolverComparisonReport'
      new_result: {}
      old_result: {}
      session_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  api.SystemInfoResponse:
    properties:
      supported_calculations:
        items:
          type: string
        type: array
      total_calculators:
        type: integer
      version:
        type: string
    type: object
  models.CalculationRequest:
    properties:
      calculation:
        description: 璁＄畻绫诲瀷
        type: string
      day:
        description: 鏃ユ湡
        type: integer
      days_from_term:
        description: 璺濊妭姘斿ぉ鏁?
        type: integer
      hour:
        description: 灏忔椂
        type: integer
      minute:
        description: 鍒嗛挓
        type: integer
      month:
        description: 鏈堜唤
        type: integer
      params:
        description: 璁＄畻鍙傛暟锛堟牸寮?锛?
      second:
        description: 绉?
        type: integer
      solar_date:
        description: 闃冲巻鏃ユ湡
        type: string
      target_longitude:
        description: 鐩爣榛勭粡
        type: number
      term_index:
        description: 鑺傛皵绱㈠紩 (0-23)
        type: integer
      year:
        description: 鐩存帴鍙傛暟锛堟牸寮?锛?
        type: integer
    required:
    - calculation
    type: object
host: localhost:8080
info:
  contact: {}
  description: 功能丰富的Go语言科学计算库，提供天文计算、历法转换、方程求解等多种科学计算功能。
  title: 科学计算工具集 API
  version: "1.0"
paths:
  /api/calculate:
    post:
      consumes:
      - application/json
      description: 执行各种类型的科学计算
      parameters:
      - description: 会话ID，用于保持计算参数一致性，不传则自动生成
        in: query
        name: session_id
        type: string
      - description: 计算请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CalculationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CalculationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 执行科学计算
      tags:
      - 科学计算
  /api/calculate-fixed:
    post:
      consumes:
      - application/json
      description: 执行各种类型的科学计算（修复版）
      parameters:
      - description: 会话ID，用于保持计算参数一致性，不传则自动生成
        in: query
        name: session_id
        type: string
      - description: 计算请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CalculationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CalculationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 执行修复版科学计算
      tags:
      - 科学计算
  /api/calculator-info:
    get:
      consumes:
      - application/json
      description: 获取指定计算器的详细信息
      parameters:
      - description: '计算类型: solar_term, ganzhi, astronomy, starting_age, lunar, planet,
          star, sunrise_sunset, moon_phase'
        in: query
        name: calculation
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CalculatorInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 获取计算器信息
      tags:
      - 计算器管理
  /api/health:
    get:
      consumes:
      - application/json
      description: 健康检查接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HealthResponse'
      summary: 健康检查
      tags:
      - 系统
  /api/solver/compare:
    post:
      consumes:
      - application/json
      description: 对比新旧方程求解器的计算结果差异
      parameters:
      - description: 计算请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CalculationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SolverComparisonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 对比新旧求解器
      tags:
      - 科学计算
  /api/system-info:
    get:
      consumes:
      - application/json
      description: 获取系统支持的完整信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SystemInfoResponse'
      summary: 获取系统信息
      tags:
      - 系统
swagger: "2.0"